requirements.txt
coverage==4.5.3
Django==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.9.4
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
